Optimization

+++ src/components/header/Header.js
@@ -1,29 +1,29 @@
-import React from 'react'
-import { HeaderFixtures } from '../../fixtures/header/HeaderFixtures'
-import SigninContainer from '../signin/SigninContainer'
+import React from "react";
+import { HeaderFixtures } from "../../fixtures/header/HeaderFixtures";
+import SigninContainer from "../signin/SigninContainer";
 
 export default function Header() {
-  const [showSignin, setShowSignin] = React.useState(false)
+  const [showSignin, setShowSignin] = React.useState(false);
   const renderClassNames = (item) => {
     if (item.route == window.location.hash) {
-      return 'nav__link active-link'
+      return "nav__link active-link";
     } else {
-      return 'nav__link'
+      return "nav__link";
     }
-  }
+  };
   React.useEffect(() => {
-    if (window.location.hash == '') {
-      window.location.hash = '#home'
-      window.scrollTo(1, 1)
+    if (window.location.hash == "") {
+      window.location.hash = "#home";
+      window.scrollTo(1, 1);
     }
-  })
+  });
   return (
     <>
       <SigninContainer
-        showSignin={showSignin && 'true'}
+        showSignin={showSignin && "true"}
         handle={() => {
-          setShowSignin(!showSignin)
-          document.body.style.overflowY = 'visible'
+          setShowSignin(!showSignin);
+          document.body.style.overflowY = "visible";
         }}
       />
       <header class="header" id="header">
@@ -43,32 +43,22 @@ export default function Header() {
                       </a>
                     </li>
                   </>
-                )
+                );
               })}
-
-              {localStorage.getItem('logged') == 'true' ? (
-                <>
-                  <button
-                    className="nav__btn"
-                    onClick={() => {
-                      localStorage.clear()
-                      window.location.reload()
-                    }}
-                  >
-                    გასვლა
-                  </button>
-                </>
-              ) : (
-                <button
-                  className="nav__btn"
-                  onClick={() => {
-                    setShowSignin(!showSignin)
-                    document.body.style.overflowY = 'hidden'
-                  }}
-                >
-                  შესვლა
-                </button>
-              )}
+              <button
+                className="nav__btn"
+                onClick={() => {
+                  if (localStorage.getItem("logged") == "true") {
+                    localStorage.clear();
+                    window.location.reload();
+                  } else {
+                    setShowSignin(!showSignin);
+                    document.body.style.overflowY = "hidden";
+                  }
+                }}
+              >
+                {localStorage.getItem("logged") == "true" ? "გასვლა" : "შესვლა"}
+              </button>
             </ul>
 
             <div class="nav__close" id="nav-close">
@@ -82,5 +72,5 @@ export default function Header() {
         </nav>
       </header>
     </>
-  )
+  );
 }
